Move lock component v1 to KM
