+++ apps/browser/src/key-management/lock/popup/lock.component.html
