+++ apps/desktop/src/key-management/lock/components/lock.component.html
