+++ apps/web/src/app/key-management/lock/components/lock.component.html
