+++ libs/angular/src/key-management/components/lock.component.ts
